Add TODO list of actions to make with this repository
Because of smaller, more targeted nature of this code base, it should
be easy to test some of the tools that I want to look at (gcov, gprof)
with this repository.

I can also test build systems that I am interested in, because
Makefiles are sometimes a bit gross to write. Particularly if I intend
to write my own kernel in the future.